requirements.txt
pyyaml
tensorflow >= 2.2